change navitaion structure
