+++ App.js
@@ -1,23 +1,23 @@
 import React, { useEffect } from 'react';
 import { anonSignIn, googleSigninConfigure, getCurrentUser } from './src/firebase/Auth';
 import AppProvider from './src/context/provider/AppProvider';
-import TabNav from './src/navigation/TabNav';
+import MainStack from './src/navigation/stack/MainStackNavigator';
 
-function App(){
-  useEffect(() => {
-    googleSigninConfigure();
-    if (getCurrentUser() == null) {
-      anonSignIn();
-    }
-  }, []);
+function App() {
+    useEffect(() => {
+        googleSigninConfigure();
+        if (getCurrentUser() == null) {
+            anonSignIn();
+        }
+    }, []);
 
-  return(
-    <>
-    <AppProvider>
-      <TabNav></TabNav>
-    </AppProvider>
-    </>
-  )
+    return (
+        <>
+            <AppProvider>
+                <MainStack></MainStack>
+            </AppProvider>
+        </>
+    )
 }
 
 export default App;
